rename structs.h to public.h
